requirements.txt
aiohttp~=3.7.3
aiosqlite~=0.17.0
Pillow~=8.2.0
dateparser~=1.0.0
python-dotenv==0.17.1
colorama~=0.4.4
sympy==1.8
psutil~=5.8.0
webcolors~=1.11.1
discord~=1.0.1
ksoftapi==0.4.1
ruamel.yaml
setuptools==56.2.0
idna==3.1
urllib3==1.26.5
redis~=3.5.3
complexity~=0.9.1
dpy-button-utils~=1.0.1
quart~=0.15.1
git+https://github.com/hikari-py/hikari@master#egg=hikari==2.0.0.dev104
git+https://github.com/FasterSpeeding/Tanjun@v2.1.2a1